JM: refactoring into shared code. ApplicationContext issues fixed
