+++ app/build.gradle
@@ -244,7 +244,10 @@ dependencies {
     testImplementation dependencies.kotlinTestJunit
 
     androidTestImplementation dependencies.truth
-    androidTestImplementation dependencies.androidxTestCore
+ //   androidTestImplementation dependencies.androidxTestCore
+//    androidTestImplementation dependencies.androidxTestMonitor
+    androidTestImplementation 'androidx.test:core:1.1.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.0'
     androidTestImplementation dependencies.runner
     androidTestImplementation dependencies.rules
     androidTestUtil dependencies.androidTestOrchestrator
