+++ shared/src/androidTest/java/de/dreier/mytargets/shared/targets/models/WAFullTest.kt
@@ -15,11 +15,10 @@
 
 package de.dreier.mytargets.shared.targets.models
 
+import androidx.test.filters.SmallTest
 import androidx.test.platform.app.InstrumentationRegistry
-import android.support.test.filters.SmallTest
-import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.runner.AndroidJUnit4
 import de.dreier.mytargets.shared.SharedApplicationInstance
-
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
@@ -34,7 +33,7 @@ class WAFullTest {
     @Before
     @Throws(Exception::class)
     fun setUp() {
-        SharedApplicationInstance.context = InstrumentationRegistry.getContext()
+        SharedApplicationInstance.context = InstrumentationRegistry.getInstrumentation().targetContext
     }
 
     @Test
