+++ shared/src/main/java/de/dreier/mytargets/shared/targets/models/WAFull.kt
@@ -50,7 +50,7 @@ class WAFull : TargetModelBase(
                 ScoringStyle(R.string.compound_style, false, 10, 9, 9, 8, 7, 6, 5, 4, 3, 2, 1),
                 ScoringStyle(false, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1),
                 ScoringStyle(true, 5, 5, 5, 4, 4, 3, 3, 2, 2, 1, 1),
-                ScoringStyle(false, 9, 9, 9, 7, 7, 5, 5, 3, 3, 1, 1),
+                ScoringStyle(R.string.imperial_outdoor_5_zone, false, 9, 9, 9, 7, 7, 5, 5, 3, 3, 1, 1),
                 ColorScoringStyle(R.string.fcfs_color_reversed, 31, 1, 1, 2, 4, 4, 6, 6, 8, 8, 10, 10)
         ),
         diameters = listOf(
