+++ shared/src/main/res/values-de/target_faces.xml
@@ -60,4 +60,7 @@
     <string name="recurve_style_x_8">Recurve-Stil (X-8)</string>
     <string name="recurve_style_10_8">Recurve-Stil (10-8)</string>
     <string name="compound_style">Compound-Stil</string>
+    <string name="imperial_outdoor_5_zone">Imperiale Außen 5 Zone</string>
+
+
 </resources>
