+++ shared/src/main/res/values-es/target_faces.xml
@@ -60,4 +60,6 @@
     <string name="recurve_style_x_8">Estilo recurvo (X-8)</string>
     <string name="recurve_style_10_8">Estilo recurvo (10-8)</string>
     <string name="compound_style">Estilo compuesto</string>
+    <string name="imperial_outdoor_5_zone">Imperial Zona 5</string>
+
 </resources>
