+++ shared/src/main/res/values/target_faces.xml
@@ -61,4 +61,6 @@
     <string name="recurve_style_x_8">Recurve style (X–8)</string>
     <string name="recurve_style_10_8">Recurve style (10–8)</string>
     <string name="compound_style">Compound style</string>
+    <string name="imperial_outdoor_5_zone">Imperial Outdoor 5 Zone</string>
+
 </resources>
